refactor(web): decompose UploadRoster into parse/import helpers + section components (Tier-2 U12)
